chore: split login/ticket/logbook/geo/class routes out of index.ts
Pure mechanical extraction, no behavior change - moves each route
group into its own file under backend/src/routes/, following the
existing registerXRoutes(app) pattern already established by
announcements.ts/stats.ts.

- New backend/src/config.ts holds ROOT_DIR/PYTHON_BIN/AI_SERVICES_DIR/
  authRateLimiter, extracted from index.ts's module scope so route
  files can import them without a circular dependency on index.ts.
- routes/auth.ts - POST /api/auth/login, GET /api/auth/me
- routes/tickets.ts - GET/POST /api/tickets*
- routes/logbook.ts - GET /api/logbook (carries forward the #208
  authorization fix unchanged)
- routes/geo.ts - /api/states, /api/districts/by-state,
  /api/blocks/by-district, /api/schools/by-block
- routes/classes.ts - GET /api/classes

index.ts drops from 3566 to ~3350 lines with these 5 groups gone.

Verified: tsc --noEmit clean on both extraction and the trimmed
index.ts; booted the real server against a scratch MongoDB seed and
curled all 6 endpoints (login, auth/me, tickets, logbook, states,
districts/by-state, classes) - all 200, identical behavior to before
the split.

Closes #133, closes #124, closes #119, closes #121, closes #120
